# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 聆櫻聖境點歌系統 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /queue:
    get:
      operationId: listQueue
      tags: [queue]
      summary: Get current queue
      responses:
        "200":
          description: Queue list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueList"
    post:
      operationId: addToQueue
      tags: [queue]
      summary: Add a song to the queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueueItemInput"
      responses:
        "201":
          description: Added to queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueItem"
  /queue/{id}:
    delete:
      operationId: removeFromQueue
      tags: [queue]
      summary: Remove an item from the queue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed from queue
  /queue/{id}/skip:
    patch:
      operationId: skipQueueItem
      tags: [queue]
      summary: Skip a queue item (no history, no play count)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Queue item skipped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueItem"
  /queue/{id}/complete:
    patch:
      operationId: completeQueueItem
      tags: [queue]
      summary: Mark a queue item as completed (writes history)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Queue item completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueItem"
  /queue/reorder:
    patch:
      operationId: reorderQueue
      tags: [queue]
      summary: Reorder the queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueueReorder"
      responses:
        "200":
          description: Queue reordered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueList"
...
  /queue/{id}/play:
    patch:
      operationId: setQueueItemPlaying
      tags: [queue]
      summary: Mark a queue item as now playing (V2 compat)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Queue item set to playing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueItem"
  /queue/{id}/start:
    patch:
      operationId: startQueueItem
      tags: [queue]
      summary: "V3: Explicitly start singing — sets started_at timestamp for lyrics sync"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Queue item started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueItem"
...
components:
  schemas:
...
    Song:
      type: object
      required: [id, title, artist, language, playCount, categories, youtubeUrl, isPracticing, hasPitchWarning, createdAt, cooldownDays, cooldownMode]
      properties:
        id:
          type: integer
        title:
          type: string
        artist:
          type: string
        language:
          type: string
        playCount:
          type: integer
        primaryTag:
          type: ["string", "null"]
        categories:
          type: array
          items:
            type: string
        youtubeUrl:
          type: ["string", "null"]
        isPracticing:
          type: boolean
        hasPitchWarning:
          type: boolean
        status:
          type: string
        createdAt:
          type: string
        lyricsText:
          type: ["string", "null"]
        lyricsFormat:
          type: ["string", "null"]
        lyricsSource:
          type: ["string", "null"]
        lyricsUpdatedAt:
          type: ["string", "null"]
        cooldownDays:
          type: integer
        cooldownMode:
          type: string
          enum: [warn, block]
...
    QueueItem:
      type: object
      required: [id, songId, requesterName, position, status, createdAt]
      properties:
        id:
          type: integer
        songId:
          type: integer
        song:
          $ref: "#/components/schemas/Song"
        requesterName:
          type: string
        note:
          type: ["string", "null"]
        position:
          type: integer
        status:
          type: string
          enum: [waiting, playing, completed, skipped]
        createdAt:
          type: string
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        sessionId:
          type: ["integer", "null"]
    QueueItemInput:
      type: object
      required: [songId, requesterName]
      properties:
        songId:
          type: integer
        requesterName:
          type: string
        note:
          type: ["string", "null"]
          maxLength: 100
    QueueList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/QueueItem"
        total:
          type: integer
    QueueReorder:
      type: object
      required: [orderedIds]
      properties:
        orderedIds:
          type: array
          items:
            type: integer
...