# lib/api-zod/src/generated/types/queueItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 聆櫻聖境點歌系統 API
 * OpenAPI spec version: 0.1.0
 */
import type { QueueItemStatus } from './queueItemStatus';
import type { Song } from './song';

export interface QueueItem {
  id: number;
  songId: number;
  song?: Song;
  requesterName: string;
  /** @nullable */
  note?: string | null;
  position: number;
  status: QueueItemStatus;
  createdAt: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  sessionId?: number | null;
}